修复报错BUG

+++ src/main/java/org/sobadfish/skywars/room/config/WorldInfoConfig.java
@@ -183,9 +183,9 @@ public class WorldInfoConfig {
         Level level = Server.getInstance().getLevelByName(pos[3]);
 
         return new Position(
-                Integer.parseInt(pos[0]),
-                Integer.parseInt(pos[1]),
-                Integer.parseInt(pos[2]),
+                Double.parseDouble(pos[0]),
+                Double.parseDouble(pos[1]),
+                Double.parseDouble(pos[2]),
                 level
 
         );
@@ -200,9 +200,9 @@ public class WorldInfoConfig {
 //            }
 //        }
         return new Location(
-                Integer.parseInt(pos[0]),
-                Integer.parseInt(pos[1]),
-                Integer.parseInt(pos[2]),
+                Double.parseDouble(pos[0]),
+                Double.parseDouble(pos[1]),
+                Double.parseDouble(pos[2]),
                 Double.parseDouble(pos[4]),
                 0, level
